Se solucionaron problemas en el taller de Hugo y se agrego la presentacion de AWS taller 2 de German

+++ Hugo Hernan Henao/Trimestre III/Evidencias/Taller Jira/Semana 2/Validator.py
@@ -51,7 +51,7 @@ class Validator():
             if len(saldo) == 0:
                 return "El saldo está vacio"
 
-            elif int(saldo) < 50 and int(saldo) > 0:
+            elif float(saldo) < 50 and float(saldo) > 0:
                 return "El saldo debe de ser mayor a 50 pesos"
 
             else:
